refac(neovim): rewrite neovim integration from scratch
As to improve latency and memory usage, this commit will remove the neovim-lib
dependency in favor of a in-house implementation of the neovim integration
logic.

The new implementation decodes the message pack messages directly to our struct
types, instead of pasing by a intermediate Value type. The decoder also uses a
arena allocator to reduce memory usage and number of allocations. It also just
borrows the string from the message buffer, instead of allocating a new one.